Some quantization helper functions.
Two functions: The first (compute_multiplier) performs the job of
picking a suitable integer qm and n so that the product of 3
floats (that are assumed to be <= 1) will equal qm/2^n.

The second function performs the steps of the
QuantizedMultiplerLessThanOne in tensorflow. I.e., it computes first a
saturated multiplication between two int32 values (using an
int64). Then it performs a rounding arithmetic right shift.